compile regex as its own feature
