+++ slosh/Cargo.toml
@@ -2,8 +2,10 @@
 name = "slosh"
 version = "0.10.0"
 edition = "2021"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+regex = ["dep:regex", "dep:lazy_static"]
 
 [dependencies]
 unicode_reader = "1"
@@ -16,11 +18,15 @@ shell = { path = "../shell" }
 unicode-width = "0.1"
 glob = "0.3"
 
+# regex dependencies
+regex = { workspace = true, optional = true }
+lazy_static = { workspace = true, optional = true }
+
 [dev-dependencies]
-regex = "1.10.2"
-lazy_static = "1.4.0"
 tempdir = "0.3.7"
 temp-env = "0.3.6"
+regex = { workspace = true }
+lazy_static = { workspace = true }
 
 [build-dependencies]
 chrono = "0.4.7"
