+++ slosh/src/main.rs
@@ -28,6 +28,8 @@ use sl_liner::{keymap, ColorClosure, Context, Prompt};
 mod completions;
 mod config;
 pub mod debug;
+#[cfg(any(test, feature = "regex"))]
+pub mod docs;
 mod liner_rules;
 mod load_eval;
 mod shell_builtins;
@@ -298,6 +300,10 @@ fn set_builtins(env: &mut SloshVm) {
     if let Ok(dir) = env::current_dir() {
         env::set_var("PWD", dir);
     }
+    #[cfg(any(test, feature = "regex"))]
+    {
+        docs::add_builtins(env);
+    }
 }
 
 fn main() {
@@ -529,312 +535,15 @@ mod tests {
     use tempdir::TempDir;
 
     use crate::tests::utils::exec;
-    use compile_state::state::{new_slosh_vm, CompileState, SloshVm, SloshVmTrait};
-    use lazy_static::lazy_static;
-    use regex::{Regex, RegexBuilder};
+    use compile_state::state::{CompileState, SloshVm, SloshVmTrait};
     use sl_compiler::pass1::pass1;
     use sl_compiler::{compile, ReadError, Reader};
     use slvm::{from_i56, Value, RET};
-    use std::borrow::Cow;
-    use std::cmp::Ordering;
-    use std::collections::HashSet;
-    use std::error::Error;
-    use std::fmt::{Debug, Display, Formatter};
-    use std::sync::Arc;
 
     // I didn't really know this was possible but for test utilities just use
     // the path attribute!
     #[path = "../../../compiler/src/test_utils/utils.rs"]
     mod utils;
-
-    lazy_static! {
-        static ref DOC_REGEX: Regex =
-            RegexBuilder::new(r#"Usage:(.*)\n\n(.*)^Section:(.+?)$(\n\n^Example:\n(.*)|\s*)"#)
-                .multi_line(true)
-                .dot_matches_new_line(true)
-                .crlf(true)
-                .build()
-                .unwrap();
-        static ref EXEMPTIONS: HashSet<&'static str> = {
-            let mut exemption_set = HashSet::new();
-            exemption_set.insert("version");
-            exemption_set.insert("env");
-            exemption_set.insert("sh");
-            exemption_set.insert("$sh");
-            exemption_set.insert("this-fn");
-            exemption_set.insert("cons");
-            exemption_set.insert("list-append");
-            exemption_set.insert("/=");
-            exemption_set.insert("eq?");
-            exemption_set.insert("equal?");
-            exemption_set.insert("type");
-            exemption_set.insert("err");
-            exemption_set.insert("call/cc");
-            exemption_set.insert("defer");
-            exemption_set.insert("on-error");
-            exemption_set.insert("while");
-            exemption_set.insert("doc-string");
-            exemption_set.insert("get");
-            exemption_set.insert("mk-err");
-            exemption_set.insert("err?");
-            exemption_set.insert("ok?");
-            exemption_set.insert("return");
-            exemption_set.insert("*euid*");
-            exemption_set.insert("*last-status*");
-            exemption_set.insert("set-prop");
-            exemption_set.insert("sizeof-heap-object");
-            exemption_set.insert("*int-min*");
-            exemption_set.insert("gensym");
-            exemption_set.insert("*uid*");
-            exemption_set.insert("*int-max*");
-            exemption_set.insert("prn");
-            exemption_set.insert("pr");
-            exemption_set.insert("sizeof-value");
-            exemption_set.insert("dump-regs");
-            exemption_set.insert("dasm");
-            exemption_set.insert("load");
-            exemption_set.insert("eval");
-            exemption_set.insert("*int-bits*");
-            exemption_set.insert("get-prop");
-            exemption_set.insert("expand-macro");
-            exemption_set
-        };
-    }
-
-    #[derive(Debug, Clone, Eq, Hash, PartialEq)]
-    enum Namespace {
-        Global,
-        // Can be adapted when namespaces are added.
-        // Other(String),
-    }
-
-    impl ToString for Namespace {
-        fn to_string(&self) -> String {
-            match self {
-                Namespace::Global => "global".to_string(),
-            }
-        }
-    }
-
-    impl Namespace {
-        fn add_docs(&self, docs: &mut Vec<SloshDoc>, vm: &mut SloshVm) -> DocResult<()> {
-            let docstring_key = vm.intern_static("doc-string");
-            match self {
-                Namespace::Global => {
-                    for g in vm.globals().keys() {
-                        let sym = Value::Symbol(*g);
-                        let sym_str = sym.display_value(&vm);
-                        let slot = vm.global_intern_slot(*g).unwrap();
-                        let raw_doc_string = vm
-                            .get_global_property(slot, docstring_key)
-                            .map_or(None, |x| {
-                                if let Value::String(h) = x {
-                                    Some(vm.get_string(h).to_string())
-                                } else {
-                                    None
-                                }
-                            })
-                            .unwrap_or_default();
-                        let slosh_doc = SloshDoc::new(
-                            sym_str,
-                            sym.display_type(&vm).to_string(),
-                            self.clone(),
-                            raw_doc_string,
-                        );
-                        match slosh_doc {
-                            Ok(slosh_doc) => {
-                                docs.push(slosh_doc);
-                            }
-                            Err(e) => match e {
-                                DocError::ExemptFromProperDocString { symbol } => {
-                                    eprintln!("Exempt from proper doc string: {symbol}");
-                                }
-                                _ => {
-                                    return Err(e);
-                                }
-                            },
-                        }
-                    }
-                }
-            }
-            docs.sort();
-            Ok(())
-        }
-    }
-
-    #[derive(Debug, Clone, Eq, Hash, PartialEq)]
-    struct DocStringSection {
-        usage: String,
-        description: String,
-        section: String,
-        example: Option<String>,
-    }
-
-    impl Display for DocStringSection {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            let example = self.example.clone().unwrap_or_default();
-            write!(
-                f,
-                "Usage: {usage}\n\n{description}\n\nSection: {section}\n\nExample: {example}",
-                usage = self.usage,
-                description = self.description,
-                section = self.section,
-                example = example,
-            )
-        }
-    }
-
-    #[derive(Eq)]
-    #[allow(unused)]
-    struct SloshDoc {
-        symbol: String,
-        symbol_type: String,
-        namespace: Namespace,
-        doc_string: DocStringSection,
-    }
-
-    impl PartialEq for SloshDoc {
-        fn eq(&self, other: &Self) -> bool {
-            self.fully_qualified_name()
-                .eq_ignore_ascii_case(&other.fully_qualified_name())
-        }
-    }
-
-    impl PartialOrd for SloshDoc {
-        fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-            self.fully_qualified_name()
-                .partial_cmp(&other.fully_qualified_name())
-        }
-    }
-
-    impl Ord for SloshDoc {
-        fn cmp(&self, other: &Self) -> Ordering {
-            self.fully_qualified_name()
-                .cmp(&other.fully_qualified_name())
-        }
-    }
-
-    impl SloshDoc {
-        fn new(
-            name: String,
-            symbol_type: String,
-            namespace: Namespace,
-            raw_doc_string: String,
-        ) -> DocResult<SloshDoc> {
-            let doc_string = SloshDoc::parse_doc_string(Cow::Borrowed(&name), raw_doc_string)?;
-            Ok(SloshDoc {
-                symbol: name,
-                symbol_type,
-                namespace,
-                doc_string,
-            })
-        }
-
-        /// Provide the fully
-        pub fn fully_qualified_name(&self) -> String {
-            self.namespace.to_string() + "::" + self.symbol.as_ref()
-        }
-
-        /// Given the rules for parsing slosh docstrings, parse one! See [`DOC_REGEX`]
-        /// for the specification.
-        pub fn parse_doc_string(
-            symbol: Cow<'_, String>,
-            raw_doc_string: String,
-        ) -> DocResult<DocStringSection> {
-            let cap = DOC_REGEX.captures(raw_doc_string.as_str()).ok_or_else(|| {
-                if EXEMPTIONS.contains(symbol.as_str()) {
-                    DocError::ExemptFromProperDocString {
-                        symbol: symbol.to_owned().to_string(),
-                    }
-                } else {
-                    DocError::NoDocString {
-                        symbol: symbol.to_owned().to_string(),
-                    }
-                }
-            })?;
-            let usage = cap
-                .get(1)
-                .ok_or_else(|| DocError::DocStringMustStartWithUsage {
-                    symbol: symbol.to_owned().to_string(),
-                })
-                .map(|x| x.as_str().to_string())?;
-            let description = cap
-                .get(2)
-                .ok_or_else(|| DocError::DocStringMissingSection {
-                    symbol: symbol.to_owned().to_string(),
-                    section: "Description".to_string(),
-                })
-                .map(|x| x.as_str().to_string())?;
-            let section = cap
-                .get(3)
-                .ok_or_else(|| DocError::DocStringMissingSection {
-                    symbol: symbol.to_owned().to_string(),
-                    section: "Section".to_string(),
-                })
-                .map(|x| x.as_str().to_string())?;
-            let example = cap.get(5).map(|x| x.as_str().to_string());
-
-            Ok(DocStringSection {
-                usage,
-                description,
-                section,
-                example,
-            })
-        }
-    }
-
-    enum DocError {
-        NoDocString { symbol: String },
-        DocStringMissingSection { symbol: String, section: String },
-        DocStringMustStartWithUsage { symbol: String },
-        ExemptFromProperDocString { symbol: String },
-    }
-
-    impl Debug for DocError {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            Display::fmt(self, f)
-        }
-    }
-
-    impl Display for DocError {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            let str = match self {
-                DocError::NoDocString { symbol } => {
-                    format!(
-                        "Either documentation provided does not conform to conventional layout or no documentation string provided for symbol {symbol} all slosh functions written in Rust must have a valid documentation string."
-                    )
-                }
-                DocError::ExemptFromProperDocString { symbol } => {
-                    format!(
-                        "No documentation needed for provided symbol {symbol}."
-                    )
-                }
-                DocError::DocStringMissingSection { symbol, section } => {
-                    format!("Invalid documentation string for symbol {symbol}, missing required section {section:?}")
-                }
-                DocError::DocStringMustStartWithUsage { symbol } => {
-                    format!(
-                        "Invalid documentation string for symbol {symbol}, first line must start with \"Usage:\""
-                    )
-                }
-            }
-            .to_string();
-            write!(f, "{}", str)
-        }
-    }
-
-    impl Error for DocError {}
-
-    type DocResult<T> = Result<T, DocError>;
-
-    #[test]
-    fn test_global_slosh_docs_formatted_properly() {
-        let mut env = new_slosh_vm();
-        set_builtins(&mut env);
-
-        let mut docs: Vec<SloshDoc> = vec![];
-        Namespace::Global.add_docs(&mut docs, &mut env).unwrap();
-    }
 
     #[test]
     fn test_load_path_no_home() {
